# hyperf/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "hyperf swagger api desc",
        "version": "1.0.0",
        "title": "HYPERF API DOC"
    },
    "host": "39.105.46.139:9503",
    "schemes": [
        "http"
    ],
...
    "paths": {
...
        "/order/get_insured_config": {
            "post": {
                "tags": [
                    "订单管理"
                ],
                "summary": "",
                "description": "获取配置",
                "operationId": "OrderGet_insured_configPOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "start_month",
                        "description": "开始月份",
                        "required": true,
                        "type": "string",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/order/add": {
            "post": {
                "tags": [
                    "订单管理"
                ],
                "summary": "",
                "description": "订单添加",
                "operationId": "OrderAddPOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "social_security_base_price",
                        "description": "社保基础价格",
                        "required": true,
                        "type": "integer",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "is_repair",
                        "description": "是否补缴",
                        "required": true,
                        "type": "integer",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "insured_id",
                        "description": "代缴人",
                        "required": true,
                        "type": "integer",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "start_month",
                        "description": "开始月份",
                        "required": true,
                        "type": "string",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "months",
                        "description": "总共月份",
                        "required": true,
                        "type": "integer",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "is_fund",
                        "description": "是否缴纳公积金",
                        "required": true,
                        "type": "integer",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "fund_base_price",
                        "description": "公积金基础价格",
                        "required": true,
                        "type": "integer",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "is_create",
                        "description": "是否生成订单",
                        "required": true,
                        "type": "integer",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "insured_area_id",
                        "description": "310000:主城区,311100:杭州余杭区",
                        "required": true,
                        "type": "integer",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/order/lists": {
            "post": {
                "tags": [
                    "订单管理"
                ],
                "summary": "",
                "description": "订单列表",
                "operationId": "OrderListsPOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "status",
                        "description": "订单状态",
                        "required": false,
                        "type": "string",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/order/detail": {
            "post": {
                "tags": [
                    "订单管理"
                ],
                "summary": "",
                "description": "订单详情",
                "operationId": "OrderDetailPOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "id",
                        "description": "订单id",
                        "required": true,
                        "type": "integer",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
...
    }
}